% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuousPvals.R
\name{continuousPvals}
\alias{continuousPvals}
\title{Correlations between each pair of continuous variables}
\usage{
continuousPvals(inputData)
}
\arguments{
\item{inputData}{Dataset that you want to sample from}
}
\value{
correlations and p-values for each pair of continuous variables
}
\description{
Comparing all combinations of continuous variables.
User can input any dataset, and this will return Pearson correlations, 95% CI, and p-value for each pair of
continuous variables
}
\examples{
continuousPvals()
}
