% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleUnivariate.R
\name{sampleUnivariate}
\alias{sampleUnivariate}
\title{Univariate sampling of variables through a dataset}
\usage{
sampleUnivariate(inputData, n, dateFormat = "\%Y\%m\%d")
}
\arguments{
\item{inputData}{dataset that you want to sample from}

\item{n}{number of rows in output/simulated dataset}

\item{dateFormat}{format of the date variable(s) (if one exists) in the original dataset. Default is "YYYYMMDD."
A set of formats can be input with c().}
}
\value{
A dataset with n rows that is distributionally similar to input dataset.
}
\description{
Takes any dataset, checks format of each variable and based on distributions of the original variables
randomly samples into a new dataset n times.
}
\details{
Returns distributionally similar dataset with user-specified number of rows
}
\examples{
sampleUnivariate()
}
